fix useServers bug, move the hook
