+++ src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useCallout } from './useCallout';
 export { default as useLocationReset } from './useLocationReset';
+export { default as useServers } from './useServers';
