--- src/settings/hooks/index.js
@@ -1,1 +1,0 @@
-export { default as useServers } from './useServers';
